+++ nextnet/tests/utility.cpp
@@ -162,7 +162,7 @@ TEST_CASE("integer_set draw_absent", "[utility]") {
 	const double M = N / (K + 1 - s.size());
 	const double SD = sqrt(M);
 	for(std::size_t i=0; i < counts.size(); ++i) {
-		if (s.find(i) == s.end())
+		if (s.find((int)i) == s.end())
 			REQUIRE(ztest(counts[i], SD, M) >= 0.01);
 		else
 			REQUIRE(counts[i] == 0);
